requirements.txt
asgiref==3.9.1
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
distlib==0.4.0
Django==5.2.4
djangorestframework==3.16.1
dnspython==2.7.0
filelock==3.19.1
Flask==3.1.2
fonttools==4.59.2
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.5
mysql-connector==2.2.9
mysql-connector-python==9.3.0
numpy==2.3.2
packaging==25.0
pandas==2.3.2
pillow==11.3.0
platformdirs==4.4.0
pymongo==4.13.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.34.0
Werkzeug==3.1.3
whitenoise==6.11.0
